// Web/src/app/Models/TermCourseOfferingModel.ts
export interface TermCourseOfferingModel {
  id: string;
  academicTermCode: string;
  academicTerm: any;
  batchCode: string;
  entry: any;
  courseCode: string;
  couse: any;
  approvedDate: string;
  staffID: string;
  staff: any;
  remark: string;
  createdDate: Date;
  lastModifiedDate: Date;
  lastModifiedBy: string;
  createdBy: string;
  isDeleted: boolean;
  academicTermSeason: number;
  academicTermYear: number;
  staffName: string;
  year: number;
  registrationEndDate: string;
  registrationStartDate: string;
  
}